+++ specutils/io/default_loaders/jwst_reader.py
@@ -141,7 +141,7 @@ def _jwst_x1d_loader(file_obj, **kwargs):
         primary_header = hdulist["PRIMARY"].header
 
         for hdu in hdulist:
-            # Read only the BinaryTableHDUs named EXTRACT1D
+            # Read only the BinaryTableHDUs named EXTRACT1D and SCI
             if hdu.name != 'EXTRACT1D':
                 continue
 
@@ -150,11 +150,14 @@ def _jwst_x1d_loader(file_obj, **kwargs):
             wavelength = Quantity(data["WAVELENGTH"])
 
             # Determine if FLUX or SURF_BRIGHT column should be returned
-            # based on whether it is point or extended source
-            srctype = hdu.header.get("srctype")
-            # SRCTYPE is in primary header because there is only one spectrum
-            if srctype is None:
-                srctype = hdulist["PRIMARY"].header.get("srctype")
+            # based on whether it is point or extended source.
+            #
+            # SRCTYPE used to be in primary header, but was moved around. As
+            # per most recent pipeline definition, it should be in the
+            # EXTRACT1D extension.
+            srctype = None
+            if "srctype" in hdu.header:
+                srctype = hdu.header.get("srctype")
 
             if srctype == "POINT":
                 flux = Quantity(data["FLUX"])